Seperated UserRequest related logic to RequestLogic. Request logic depends on WordLogic for validating requests
