+++ application/dict-app/app/objects/Word.java
@@ -3,6 +3,7 @@ package objects;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.google.common.base.Preconditions;
 import lombok.*;
 import utilities.JsonUtil;
